pyproject.toml
[tool.poetry]
name = "mlops2023"
version = "0.1.0"
description = ""
authors = ["german <mr.g.o@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.7.0,<3.11"
pandas = "^1.2.2"
sklearn = "0.0.post9"
joblib = "^1.3.2"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"
scipy = "1.7.3"
scikit-learn = "^0.23"

[tool.poetry.group.dev.dependencies]
dvc-gdrive = {version = "^2.20.0", python = "^3.8"}
dvc = "^1.11.6"
click = "^8.1.7"

[tool.black]
line-length = 90
target-version = ["py37"]
skip-string-normalization = true

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 90
lines_after_imports = 2

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
